+++ src/test/java/nextstep/domain/ReservationTest.java
@@ -1,4 +1,4 @@
-package nextstep;
+package nextstep.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
